pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mediadive-growth-db"
version = "0.1.0"
description = "ML pipeline for predicting and generating viable microbial growth media from metagenomics data"
readme = "readme.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "MediaDive Research Team"},
]

dependencies = [
    "requests>=2.31",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
ml = [
    "numpy>=1.26",
    "pandas>=2.1",
    "scikit-learn>=1.4",
    "scipy>=1.12",
    "torch>=2.2",
    "lightning>=2.2",
    "transformers>=4.38",
    "datasets>=2.17",
    "wandb>=0.16",
]
bio = [
    "biopython>=1.83",
    "rdkit",
]
viz = [
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "plotly>=5.18",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "pytest-mock>=3.12",
    "ruff>=0.3",
    "mypy>=1.8",
    "pre-commit>=3.6",
    "ipykernel>=6.29",
    "jupyter>=1.0",
    "nbstripout>=0.7",
]
cloud = [
    "boto3>=1.34",
    "modal>=0.62",
]
all = [
    "mediadive-growth-db[ml,bio,viz,dev,cloud]",
]

[tool.setuptools.packages.find]
include = ["src*"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests requiring network/DB access",
    "unit: marks pure unit tests",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
